requirements.txt
PyYAML>=6.0

# Optional / testing dependencies (uncomment to install)
# pytest>=7.0
